PythonProject4/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
django-filter>=24.1
psycopg2-binary==2.9.9
